# apps/frontend/public/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "ArcPay Mantle Developer API",
    "version": "0.1.2",
    "summary": "Public developer surface for ArcPay's Mantle x402, workspace evidence, and agent integration flows.",
    "description": "ArcPay Mantle exposes a wallet-first app, a live x402 payment server, an MCP server for local agent hosts, and OpenAPI-compatible endpoint docs for developers building on Mantle Testnet."
  },
  "servers": [
    {
      "url": "https://mantle-x402.20.208.46.195.nip.io",
      "description": "Live ArcPay Mantle x402 server"
    },
    {
      "url": "https://arcpay-mantle.vercel.app",
      "description": "ArcPay Mantle frontend and workspace API"
    }
  ],
...
  "paths": {
...
    "/x402/payment-requirements/{agentSlug}": {
      "get": {
        "tags": ["x402"],
        "summary": "Quote Mantle payment requirements",
        "description": "Returns exact MNT payment requirements for a registered Mantle agent without requesting the protected resource.",
        "operationId": "getPaymentRequirements",
        "parameters": [
          {
            "name": "agentSlug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "example": "research-agent"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Payment requirements for the agent.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaymentRequirements"
                }
              }
            }
          },
          "404": {
            "description": "Agent not found or inactive."
          }
        }
      }
    },
    "/agent/{agentSlug}/work": {
      "get": {
        "tags": ["x402"],
        "summary": "Request protected agent work",
        "description": "Returns HTTP 402 with x402-style payment requirements until a supplied Mantle order proof is fulfilled or settled.",
        "operationId": "getProtectedAgentWork",
        "parameters": [
          {
            "name": "agentSlug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "example": "research-agent"
            }
          },
          {
            "name": "orderId",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "example": "0xabc123..."
            }
          },
          {
            "name": "X-Payment",
            "in": "header",
            "required": false,
            "description": "Mantle payment proof. Accepts a raw orderId, JSON like {\"orderId\":\"0x...\"}, or base64url JSON.",
            "schema": {
              "type": "string",
              "example": "{\"orderId\":\"0xabc123...\"}"
            }
          },
          {
            "name": "X-ArcPay-Order-Id",
            "in": "header",
            "required": false,
            "description": "Direct ArcPay order proof header for agents that do not yet produce X-Payment payloads.",
            "schema": {
              "type": "string",
              "example": "0xabc123..."
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Unlocked agent result.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnlockedWork"
                }
              }
            }
          },
          "402": {
            "description": "Payment required.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaymentRequirements"
                }
              }
            }
          }
        }
      }
    },
    "/x402/verify": {
      "post": {
        "tags": ["x402"],
        "summary": "Verify an on-chain order",
        "description": "Checks AgentOrderBook state and returns whether the order is paid, fulfilled, unlocked, or settled.",
        "operationId": "verifyX402Order",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VerifyOrderRequest"
              },
              "example": {
                "orderId": "0xabc123...",
                "agentSlug": "research-agent"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Verification result.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VerificationResult"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "adminBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    },
    "schemas": {
...
      "PaymentRequirements": {
        "type": "object",
        "properties": {
          "ok": { "type": "boolean" },
          "x402Version": { "type": "string" },
          "arcpayScheme": { "type": "string" },
          "protocol": { "type": "string", "example": "x402" },
          "network": { "type": "string" },
          "caip2Network": { "type": "string", "example": "eip155:5003" },
          "chainId": { "type": "integer" },
          "currency": { "type": "string" },
          "resource": { "type": "string" },
          "agent": { "$ref": "#/components/schemas/Agent" },
          "accepts": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/AcceptRequirement" }
          }
        }
      },
      "Agent": {
        "type": "object",
        "properties": {
          "slug": { "type": "string" },
          "agentId": { "type": "string" },
          "owner": { "type": "string" },
          "name": { "type": "string" },
          "endpoint": { "type": "string" },
          "capabilities": { "type": "string" }
        }
      },
      "AcceptRequirement": {
        "type": "object",
        "properties": {
          "scheme": { "type": "string", "example": "exact" },
          "network": { "type": "string", "example": "eip155:5003" },
          "chainId": { "type": "integer", "example": 5003 },
          "asset": { "type": "string", "example": "native" },
          "currency": { "type": "string", "example": "MNT" },
          "maxAmountRequired": { "type": "string" },
          "amountWei": { "type": "string" },
          "amountStt": { "type": "string" },
          "payTo": { "type": "string" },
          "recipient": { "type": "string" },
          "description": { "type": "string" },
          "resource": { "type": "string" },
          "mimeType": { "type": "string" },
          "contract": { "type": "string" },
          "action": { "type": "string" },
          "calldata": { "type": "string" },
          "args": {
            "type": "object",
            "properties": {
              "agentId": { "type": "string" },
              "requestUri": { "type": "string" }
            }
          },
          "unlockUrl": { "type": "string" },
          "verificationUrl": { "type": "string" },
          "paymentProof": {
            "type": "object",
            "properties": {
              "type": { "type": "string", "example": "mantle-order-id" },
              "header": { "type": "string", "example": "X-Payment" },
              "acceptedFormats": {
                "type": "array",
                "items": { "type": "string" }
              }
            }
          }
        }
      },
      "VerifyOrderRequest": {
        "type": "object",
        "required": ["orderId"],
        "properties": {
          "orderId": { "type": "string" },
          "agentSlug": { "type": "string" },
          "requester": { "type": "string" }
        }
      },
      "VerificationResult": {
        "type": "object",
        "properties": {
          "ok": { "type": "boolean" },
          "paid": { "type": "boolean" },
          "unlocked": { "type": "boolean" },
          "settled": { "type": "boolean" },
          "statusName": { "type": "string" },
          "amountStt": { "type": "string" },
          "requester": { "type": "string" },
          "provider": { "type": "string" },
          "resultUri": { "type": "string" }
        }
      },
      "UnlockedWork": {
        "type": "object",
        "properties": {
          "ok": { "type": "boolean" },
          "unlocked": { "type": "boolean" },
          "result": {
            "type": "object",
            "properties": {
              "summary": { "type": "string" },
              "evidenceUri": { "type": "string" },
              "generatedAt": { "type": "string" },
              "nextAction": { "type": "string" }
            }
          }
        }
      },
...
    }
  }
}